// src/webgl/values/enums.ts
export enum ComponentTypeEnum {
  CHILDREN = 'children',
  STYLE = 'style',
}

export enum NavigationOrientation {
  HORIZONTAL = 'horizontal',
  VERTICAL = 'vertical',
}

export enum NavigationDirection {
  LEFT = 'left',
  RIGHT = 'right',
  UP = 'up',
  DOWN = 'down',
}

export enum NavigationDirectionEvent {
  LEFT = 'ArrowLeft',
  RIGHT = 'ArrowRight',
  UP = 'ArrowUp',
  DOWN = 'ArrowDown',
}

export enum NavigationResponsibility {
  ADD = 'add',
  SUBTRACT = 'subtract',
}